Exam is done!
The task implemented a connection to the "rickandmortyapi" API and getting information about the characters and the episodes in which they appeared.

When you click on the "Search" button, a check will be made for emptiness and the correctness of the text input, if emptiness was entered, not a number, an incorrect number, or a character with the same id has already been added, the user will receive a corresponding error. After a successful check, the page will display the characters whose id the user has entered.

When you click on the "Delete" button, the selected character is removed from the page. When clicking on a picture with a character, the user is taken to a page with detailed information about the character and the episodes in which he appeared.

The user has the option to return to the home page by clicking on the "Home" button. After returning to the main page, all the characters that were previously uploaded by the user are displayed.